requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-appconf==1.0.2
django-crispy-forms==1.6.1
django-geoip==0.5.2.1
djangorestframework==3.7.7
geoip2==2.5.0
idna==2.6
maxminddb==1.3.0
pytz==2017.2
requests==2.18.4
urllib3==1.22
